Support customize gorm field type
